pyproject.toml
[tool.poetry]
name = "tree-everywhere"
version = "0.1.0"
description = "A database of trees planted by volunteers around the world."
authors = ["Vinicius Peixoto <vinicius.opeixoto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dynaconf = {extras = ["yaml"], version = "^3.2.5"}
djangorestframework = "^3.15.2"
uvicorn = "^0.30.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
blue = "^0.9.1"
ruff = "^0.4.10"
isort = "^5.13.2"

[tool.ruff]
line-length = 100
exclude = ["migrations"]

[tool.isort]
profile = "black"
line_length = 100
extend_skip = ["migrations"]
split_on_trailing_comma = "true"

[tool.pytest.ini_options]
pythonpath = "."
DJANGO_SETTINGS_MODULE = "psys.settings"

[tool.blue]
extend-exclude = """
(
    ^/manage.py
    | migrations/
)
"""

[tool.taskipy.tasks]
lint = 'ruff check . && blue --check . --diff && isort --check . --diff'
format = 'blue . && isort .'
run = 'uvicorn --reload psys.asgi:application'
pre_test = 'task lint'
test = 'pytest -s -x --cov=trees --cov-config=".coveragerc" -vv'
post_test = 'coverage html -d trees_html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
